refactor: update constants and import paths according to repackaging
