+++ data/src/main/java/com/london/data/mapper/details/ImagesMapper.kt
@@ -1,6 +1,6 @@
 package com.london.data.mapper.details
 
-import com.london.data.remote.model.details.ImagesResponse
+import com.london.data.remote.model.details.image.ImagesResponse
 import com.london.data.utils.asImageUrlOrEmpty
 import com.london.data.utils.orZero
 import com.london.domain.entity.shared.ImagesEntity
