+++ data/src/main/java/com/london/data/mapper/details/actor/ActorTvShowDetailsMapper.kt
@@ -7,8 +7,8 @@ import com.london.data.remote.model.details.actor.tvshow.ActorTvShowDetailsRespo
 import com.london.data.utils.asImageUrlOrEmpty
 import com.london.data.utils.orZero
 import com.london.domain.KoverIgnore
-import com.london.domain.entity.actor.cast.ActorMediaDetails
-import com.london.domain.entity.actor.cast.ActorMediaItems
+import com.london.domain.entity.actor.ActorMediaDetails
+import com.london.domain.entity.actor.ActorMediaItems
 
 fun ActorTvShowDetailsResponse.toEntity(): ActorMediaDetails =
     ActorMediaDetails(
