+++ data/src/main/java/com/london/data/remote/source/tvshow/TvShowRemoteDataSource.kt
@@ -2,8 +2,8 @@ package com.london.data.remote.source.tvshow
 
 import com.london.data.remote.model.ApiResponse
 import com.london.data.remote.model.account.AccountStatesResponse
-import com.london.data.remote.model.details.ImagesResponse
 import com.london.data.remote.model.details.actor.tvshow.ActorTvShowDetailsResponse
+import com.london.data.remote.model.details.image.ImagesResponse
 import com.london.data.remote.model.details.rating.RatingRemoteResponse
 import com.london.data.remote.model.details.tvshow.TvShowDetailsRemoteResponse
 import com.london.data.remote.model.details.tvshow.episode.EpisodeDetailsResponse
