+++ domain/src/main/java/com/london/domain/entity/actor/ActorMediaDetails.kt
@@ -1,4 +1,4 @@
-package com.london.domain.entity.actor.cast
+package com.london.domain.entity.actor
 
 data class ActorMediaDetails(
     val mediaItems: List<ActorMediaItems> = emptyList(),
