+++ domain/src/main/java/com/london/domain/entity/actor/ActorMediaItems.kt
@@ -1,4 +1,4 @@
-package com.london.domain.entity.actor.cast
+package com.london.domain.entity.actor
 
 data class ActorMediaItems(
     val id: Int = 0,
