+++ domain/src/main/java/com/london/domain/entity/shared/MediaStates.kt
@@ -1,8 +1,8 @@
 package com.london.domain.entity.shared
 
 data class MediaStates(
-    val favorite: Boolean,
     val id: Int,
+    val favorite: Boolean,
     val rate: Int,
     val watchlist: Boolean
 )
